test(static embeddings): add word2vec tests
