+++ tests/test_automodel.py
@@ -4,6 +4,7 @@ from transformers import BertModel
 
 from embeddings.embedding.auto_flair import AutoFlairWordEmbedding
 from embeddings.embedding.flair_embedding import FlairTransformerWordEmbedding
+from embeddings.embedding.static.word import AutoStaticWordEmbedding
 
 
 def test_automodel_transformer() -> None:
@@ -21,5 +22,17 @@ def test_automodel_error_repo_not_found() -> None:
 def test_automodel_error_wrong_format() -> None:
     with pytest.raises(EnvironmentError):
         AutoFlairWordEmbedding.from_hub(
+            repo_id="sentence-transformers/average_word_embeddings_glove.6B.300d"
+        )
+
+
+def test_static_automodel_error_repo_not_found() -> None:
+    with pytest.raises(EnvironmentError):
+        AutoStaticWordEmbedding.from_default_config(repo_id="name_of_repo_that_does_not_exist")
+
+
+def test_static_automodel_error_wrong_format() -> None:
+    with pytest.raises(EnvironmentError):
+        AutoStaticWordEmbedding.from_default_config(
             repo_id="sentence-transformers/average_word_embeddings_glove.6B.300d"
         )
